requirements.txt
Django==3.0
gunicorn==19.5.0
slackclient>=2.0.0
certifi==0.0.8
djangorestframework